Updated controller.cpp to use devkitpro filesystem
An filesystem error was occurring when using the standard c++ filesystem library. Not sure why, but I decided to update everything to use devkitpro's filesystem library. This is what most Switch homebrew seems to be using, so it MIGHT help.

Also fixed VS Code's Intellisense issue. For some reason, it was internally trying the use Windows cpp_properties configuration even when I changed it in the settings. Re-ordering the configurations so the Linux one is first seems to address the issue.

- Changed the Makefile to allow exceptions for the time being while I'm not sure how likely the tool will work. I may want to add exception handling when debugging in the future.