Fixed failure of getting the running game's title ID
Apparently these kinds of services cause errors if started in the middle of runtime.

Going to rant again that I wish libnx was documented. It took hours of brute force to discover this was the problem.

Leaving dmntcht.h in the dependency folder just in case I need it in the future, but it will probably just be removed.

Other improvements while I was trying to fix it:

- SD card is now mounted in Controller's constructor. Not sure why I didn't do that to begin with.