Fixed InvalidMemoryAccess error
Apparently the filesystem library being used should not have "sdmc:" at the beginning of the paths. Not sure where I got that.

Also noticed the game-folder-not-found message was not being wrapped, but was being cut-off. Split it into multiple lines to help address that.

Also removed doesFileNotExist(). It's better to call !doesFileExist(). 1 less layer on the call stack.