Fixed Stack Overflow
Again, there's like no documentation for libnx. It turns out I misunderstood some of the filesystem parameters because of that lack of documentation.

It was an iterator. It would take a null s64 and put the content total into it.

I updated all the relevant filesystem libnx function code blocks to now handle the task appropriately.

I also noticed other homebrew source code initializes the folder buffer to the size of FS_MAX_PATH, so I added a function to do so.

I was also missing the setContent in ui_groups