Fixed memory access exception
Discovered the issue to the problem I was having was the [&] in the lambdas for navigating to a different UI. This was not giving me access to the needed values from the surrounded class. I think the values were being thrown out.

The linter and copilot were not able to notice that issue.

Also removed the GuiMain because I'm starting to think I can implement all the features for this overlay that I want without that menu, so this will probably eliminate overhead to some degree.