Fixed issue of fsDirRead only reading the first file
I discovered that the second parameter of fsDirRead is actually not supposed to be the total. I had it correct to begin with.

I also discovered that the second parameter will be set to 0 when it has iterated over all contents in the path. Adding '&& count' ensures the code will only run if count is not 0.

It's still a mystery to me if the 'count' iterates down to 0, or starts at 0, iterating to the total (but returning to 0 when the total has been iterated over).

Currently guessing it goes up. Activating a mod will definitely break if I'm wrong about that.