Added code to move files back and forth on start/close
WIP CURRENTLY BROKEN

Split the code for activating/deactivating mods into separate methods.

Added methods that use them for enabling/disabling mods. This is different from activating because it leaves the file in existence, marking it as active/inactive even though the files may be in the opposite location.

Also fixed broken toggles in the Mods UI (I think this time).