Fixed broken non-default trackbars
The issue turned out to be a simple misunderstanding on how `std::map`'s `insert(...)` function operated.

I expected it to work in the manner that the bracket operator behaves, so I replaced it with that.

While I was experimenting with fixing this issue, I replaced the `StepTrackBar`s with regular `TrackBar`s (requiring me to make adjustments so that the saving & loading of ratings is on a scale of 0-100 now instead of 0-10).
* Not sure which I like better, but I decided to just leave the regular `TrackBar` in.