Set version for initial alpha release (v0.1))
I originally was hoping to initially release this as a more-stable version, but I've lost confidence with its stability.

There are problematic bugs that are out of my control (or are seemingly out of my control) that I am unable to fix.

I will file Issues on these, so hopefully they can get addressed eventually, but for now, I already spent more time getting this app 95% of the way working as I should. At least it's to the point where it will be useful to many people, although its problems may get chase a minority away.